# Fixing the Performance Drop in Random Reads - Implemented using Go Storage Client