Issue:

Because of NopCloser conversion  for converting Reader object to ReadCloser, we got less bandwidth and more latencies for random reads.
For 50mb File size
Bandwidth - 224 MiB/s
Latency - 12084.52 ms

Fixes:

After rechecking http_client [NewRangeReader](https://github.com/GoogleCloudPlatform/gcsfuse/blob/master/vendor/cloud.google.com/go/storage/http_client.go#L771) method we got  to know that the object is ReadCloser type only so we don't need extra conversion of NopCloser.
For 50mb File size
Bandwidth - 1184MiB/s
Latency - 2199.80ms

Reason:

This could be because of we are copying heavy object many times during conversion.